Refactor data generator runner into exported package
