+++ .gitignore
@@ -1,3 +1,4 @@
 random_data_load*
+*.swp
 vendor/
 bin/*
